add service Categories and null image

+++ controllers/handlerFactory.js
@@ -2,13 +2,36 @@ const { Model } = require("mongoose");
 const AppError = require("../utils/appError");
 const catchAsync = require("../utils/catchAsync");
 
-exports.deleteOne = (Model) =>
+
+exports.deleteOne = (Model, relations = []) =>
   catchAsync(async (req, res, next) => {
-    const doc = await Model.findByIdAndDelete(req.params.id);
-    if (!doc) return next(new AppError("No Entitie found", 404));
+    // Find the document
+    let doc = await Model.findById(req.params.id);
+    if (!doc) return next(new AppError("No entity found", 404));
+
+    let hasRelations = false;
+
+    for (const relation of relations) {
+      const relatedDocs = await doc.populate(relation).execPopulate();
+      if (relatedDocs[relation] && relatedDocs[relation].length > 0) {
+        hasRelations = true;
+        break;
+      }
+    }
+
+  
+    if (hasRelations) {
+      doc.deletedAt = Date.now();
+      await doc.save(); 
+    } else {
+      await Model.findByIdAndDelete(req.params.id);
+    }
+
     res.status(200).json({
       status: "success",
-      message: "Entitie was Deleted Successfully",
+      message: hasRelations
+        ? "Entity was soft-deleted successfully"
+        : "Entity was deleted successfully",
     });
   });
 
@@ -21,6 +44,7 @@ exports.updateOne = (Model) =>
     } else if (req.files) {
       updates.images = req.files.map((file) => file.path);
     }
+    updates.updatedAt = Date.now();
 
     const doc = await Model.findByIdAndUpdate(req.params.id, updates, {
       new: true,
@@ -73,7 +97,7 @@ exports.getAll = (Model) =>
 
 exports.getOne = (Model) =>
   catchAsync(async (req, res, next) => {
-    const getModel = await Model.findById(req.params.id);
+    const getModel = await Model.findById({ _id: req.params.id, deletedAt: null });
     res.status(200).json({
       status: "success",
       message: "Got the Selected Entitie Successfully",
@@ -85,7 +109,7 @@ exports.getOne = (Model) =>
 
 exports.getWithFilter = (Model) =>
   catchAsync(async (req, res, next) => {
-    const filter = req.filter || {};
+    const filter = { ...req.filter, deletedAt: null };
 
     const getModels = await Model.find(filter);
     //const getModels = queryModel.exec();
